refactor(tabs): 抽取 useSyncActiveTabSideEffects 消除 close-tab 副作用重复
TabBar.handleClose 与 GlobalShortcuts.handleCloseTab 各自实现了约 30
行"关闭活跃标签后同步 appMode/currentXxxId/workspace/unviewed 标记"
的副作用，将来新增同步项容易漏一处——实际上 GlobalShortcuts 已经漏掉
了"清除 Agent 会话的已完成未查看标记"。

抽出 useSyncActiveTabSideEffects 统一封装这段逻辑，两处调用新 hook，
顺带修复 GlobalShortcuts 的漏清除问题。
